Revert "Moving functiond from tcp parser to seperate scripts"
