+++ src/ble_device_handlers/smart_bike_trainer.rs
@@ -73,7 +73,7 @@ pub async fn get_smart_trainer_device(
     Ok(BleDevice::SmartTrainer {
         control_char,
         data_char,
-        peripheral,
+        peripheral_index,
     })
 }
